requirements.txt
flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
APScheduler==3.10.4
sqlalchemy==2.0.29
pandas==2.2.2
openpyxl==3.1.2
spacy==3.8.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
cryptography==41.0.0
sentence-transformers
scikit-learn
tf-keras
streamlit
